onion/requirements.txt
requests
beautifulsoup4
tqdm
sentence_transformers
psycopg2-binary
pgvector